requirements.txt
click==8.1.2
python-dotenv==0.20.0
rich==12.2.0
tabulate==0.8.9
terra-sdk==2.0.4
